Hindsight is 20/20?
I update the version of Unity this uses to 2019 and made some adjustments to the UI to make buttons easier to read and click. 
I recorded and added in new sound effects for when Swole is hit with holy water.  I also added in a new sound effect for when the candles are placed properly.
Speaking of the candles, I created new sprites/materials for them so they are now easier to recognize using colors and shapes.  I also changed how the candle prompt was handled by making it part of the UI instead of a cube floating in midair.
I made the timing for thurible swinging slightly longer and added in text to make sure the player knows which buttons to hit when each one lights up.